Update DashUI.py - Adding dash datatable
Dash datatable is being used where the current sentence relations used to be displayed. The datatable is better for this as it is easier to read, and it is editable.
Some of the functionality:
Whenever anything is edited in the datatable (including row removal), it updates the all-relation-store to reflect that change.