Update DashUI.py - a lot changed in code
A lot changed in code due to adding dash bootstrap components.
Plus new features: Left and right arrow will now move you through the paper.
Up arrow will set direction to increase, and and down arrow will set it to decrease.
Hitting the plus key on the numpad will also set the direction to increase, and hitting the minus key on the numpad will also set the direction to decrease.